Renaming to new semantics started
