+++ src/main/java/com/artemka091102/explosion/Main.java
@@ -8,7 +8,7 @@ import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 @Mod(Main.MODID)
 public class Main {
     public Main() {
-        FMLJavaModLoadingContext.get().getModEventBus().addListener(CrackedDict::onCommonSetup);
+        FMLJavaModLoadingContext.get().getModEventBus().addListener(ModsController::onCommonSetup);
     }
 
     public static final String MODID = "explosion";
